--- members/function.json
@@ -1,21 +1,0 @@
-{
-  "bindings": [
-    {
-      "authLevel": "anonymous",
-      "type": "httpTrigger",
-      "direction": "in",
-      "name": "req",
-      "route": "signups/{signupId}/members/{memberId?}",
-      "methods": [
-        "GET",
-        "POST",
-        "PUT"
-      ]
-    },
-    {
-      "type": "http",
-      "direction": "out",
-      "name": "res"
-    }
-  ]
-}
